Release 0.0.1.23 stable interaction checkpoint

+++ src/game/Tactical/OS0_RealtimeEditor.cc
@@ -1360,7 +1360,7 @@ try
 			basic.bRelativeEquipmentLevel = std::clamp<INT8>(
 				request.relativeEquipmentLevel, 0, 4);
 			basic.bDirection = request.direction >= 0 && request.direction < 8 ?
-				request.direction : NORTHWEST;
+				request.direction : static_cast<INT8>(NORTHWEST);
 			basic.bOrders = found->orders;
 			basic.bAttitude = found->attitude;
 			basic.bBodyType = found->bodyType;
